requirements.txt
langchain==0.3.13
langchain-openai==0.2.9
streamlit==1.41.1
python-dotenv==1.0.1
loguru==0.7.2
watchdog==4.0.0
numexpr==2.10.2
docx2txt==0.8
chromadb==0.6.0
langchain_community==0.3.13
pymupdf==1.25.1
psycopg2-binary==2.9.10
langgraph==0.2.61